backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
SQLAlchemy==2.0.32